requirements.txt
aiomysql==0.1.1
alembic==1.10.2
anyio==3.6.2
boto3==1.26.100
botocore==1.29.100
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
fastapi==0.95.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
strawberry-graphql==0.168.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==10.4
